requirements.txt
absl-py==2.1.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.1
all-clip==1.2.0
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
autofaiss==2.17.0
av==11.0.0
blinker==1.7.0
blurhash==1.1.4
boost==0.1
braceexpand==0.1.7
build==1.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip-anytorch==2.6.0
clip-retrieval==2.44.0
cmake==3.28.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
dataclasses==0.6
decorator==5.1.1
decord==0.6.0
diffusers @ file:///home/ids/xchen-21/FADING/diffusers
docker-pycreds==0.4.0
easydict==1.11
einops==0.7.0
embedding-reader==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
ExifRead-nocycle==3.0.1
faiss-cpu==1.7.4
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.42
gradio==4.12.0
gradio_client==0.8.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
img2dataset==1.45.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
insightface==0.7.3
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git@305a205685d1b0257412906f2856ae699ee8202c
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
kornia==0.7.1
lazy_loader==0.3
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
Mastodon.py==1.8.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multilingual-clip==1.0.10
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.17.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prettytable==3.9.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
sentry-sdk==1.40.4
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.25.0
wandb==0.16.3
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
Werkzeug==3.0.1
xformers==0.0.23.post1
yarl==1.9.4
zipp==3.17.0
